pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "blinker>=1.9.0",
    "click>=8.2.1",
    "email-validator>=2.2.0",
    "flask-dance>=7.1.0",
    "flask>=3.1.1",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "itsdangerous>=2.2.0",
    "jinja2>=3.1.6",
    "markupsafe>=3.0.2",
    "psycopg2-binary>=2.9.10",
    "pyjwt>=2.10.1",
    "werkzeug>=3.1.3",
    "flask-login>=0.6.3",
    "oauthlib>=3.3.1",
    "sqlalchemy>=2.0.41",
    "requests>=2.32.4",
    "pyyaml>=6.0.2",
    "sendgrid>=6.12.4",
    "python-dotenv>=1.1.1",
]
